requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pdfplumber>=0.10.0
pymupdf>=1.23.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.0
mistralai>=1.0.0
